+++ apps/web/app/(org)/dashboard/_components/AnimatedIcons/Refer.tsx
@@ -53,6 +53,7 @@ const ReferIcon = forwardRef<ReferIconHandle, ReferIconProps>(
 				onAnimationComplete={() => setIsAnimating(false)}
 				{...props}
 			>
+				<title>Box</title>
 				<path
 					strokeLinecap="round"
 					strokeLinejoin="round"
