+++ apps/web/app/globals.css
@@ -52,6 +52,18 @@ html {
 	background-color: #f2f2f2;
 }
 
+
+.dashboard-grid {
+	display: grid;
+	grid-template-columns: auto 1fr 1fr;
+	grid-template-areas:
+		"sidebar main main"
+		"sidebar main main";
+	height: 100dvh;
+	width: 100vw;
+	min-height: 100dvh;
+}
+
 .left-perspective {
 	transform: scale(1.2) rotateX(47deg) rotateY(31deg) rotate(324deg);
 }
@@ -66,9 +78,14 @@ html {
 	border-radius: 10px;
 }
 
-/* Handle */
 ::-webkit-scrollbar-thumb {
-	background: rgb(116, 116, 116);
+	background: rgb(178, 178, 178);
+	border-radius: 10px;
+}
+
+/* Handle */
+.dark ::-webkit-scrollbar-thumb {
+	background: rgb(61, 61, 61);
 	border-radius: 10px;
 }
 
@@ -289,6 +306,15 @@ a.bg-primary:hover {
 		linear-gradient(#fff 0 0);
 	-webkit-mask-composite: xor;
 	mask-composite: exclude;
+}
+
+@media all and (max-width: 1024px) {
+	.dashboard-grid {
+		grid-template-columns: 1fr;
+		grid-template-areas:
+			"main"
+			"main";
+	}
 }
 
 @media (prefers-color-scheme: dark) {
